requirements.txt
python-dotenv
pandas
requests
pymongo
certifi
pymongo[srv]==3.11

-e .